+++ packages/core/src/hydrate.test.ts
@@ -1,12 +1,44 @@
-import { config } from '@wagmi/test'
+import { accounts, config, wait } from '@wagmi/test'
+import type { EIP1193Provider } from 'mipd'
 import { http } from 'viem'
 import { mainnet } from 'viem/chains'
-import { expect, test } from 'vitest'
+import { expect, test, vi } from 'vitest'
 
+import { createConnector } from './connectors/createConnector.js'
+import { mock } from './connectors/mock.js'
 import { createConfig } from './createConfig.js'
 import { createStorage } from './createStorage.js'
 import { hydrate } from './hydrate.js'
 import { cookieStorage } from './utils/cookie.js'
+
+vi.mock(import('mipd'), async (importOriginal) => {
+  const mod = await importOriginal()
+
+  let cache: typeof mod | undefined
+  if (!cache)
+    cache = {
+      ...mod,
+      createStore() {
+        const store = mod.createStore()
+        return {
+          ...store,
+          getProviders() {
+            const info = {
+              icon: 'data:image/svg+xml,<svg width="32px" height="32px" viewBox="0 0 32 32"/>',
+              uuid: crypto.randomUUID(),
+            } as const
+            const provider = '<EIP1193Provider>' as unknown as EIP1193Provider
+            return [
+              { info: { ...info, name: 'Foo', rdns: 'com.foo' }, provider },
+              { info: { ...info, name: 'Bar', rdns: 'com.bar' }, provider },
+              { info: { ...info, name: 'Mock', rdns: 'com.mock' }, provider },
+            ]
+          },
+        }
+      },
+    }
+  return cache
+})
 
 test('default', () => {
   const { onMount } = hydrate(config, {
@@ -14,6 +46,7 @@ test('default', () => {
     reconnectOnMount: false,
   })
   onMount()
+
   expect(onMount).toBeDefined()
 })
 
@@ -35,15 +68,24 @@ test('initialState', () => {
     reconnectOnMount: true,
   })
   onMount()
+
   expect(onMount).toBeDefined()
 })
 
-test('ssr', () => {
+test('ssr', async () => {
   const config = createConfig({
     chains: [mainnet],
-    transports: { [mainnet.id]: http() },
+    connectors: [
+      createConnector((c) => {
+        return {
+          ...mock({ accounts })(c),
+          rdns: 'com.mock',
+        }
+      }),
+    ],
     ssr: true,
     storage: createStorage({ storage: cookieStorage }),
+    transports: { [mainnet.id]: http() },
   })
 
   const { onMount } = hydrate(config, {
@@ -58,4 +100,15 @@ test('ssr', () => {
   onMount()
   expect(onMount).toBeDefined()
   expect(config.chains[0].id).toBe(1)
+
+  await wait(100)
+  expect(config.connectors.map((x) => x.rdns ?? x.id)).toMatchInlineSnapshot(
+    `
+    [
+      "com.mock",
+      "com.foo",
+      "com.bar",
+    ]
+  `,
+  )
 })